chore: refactor sub dir paths
