+++ packages/sdk/src/index.ts
@@ -10,9 +10,9 @@ export { utf8ToBytes, bytesToUtf8 } from "@waku/utils/bytes";
 export * from "./waku.js";
 
 export { createLightNode, defaultLibp2p } from "./create/index.js";
-export { wakuLightPush } from "./protocols/light_push.js";
-export { wakuFilter } from "./protocols/filter.js";
-export { wakuStore } from "./protocols/store.js";
+export { wakuLightPush } from "./protocols/lightpush/index.js";
+export { wakuFilter } from "./protocols/filter/index.js";
+export { wakuStore } from "./protocols/store/index.js";
 
 export * as waku from "@waku/core";
 export * as utils from "@waku/utils";
