+++ packages/sdk/src/protocols/filter/constants.ts
@@ -1,0 +1,5 @@
+export const DEFAULT_KEEP_ALIVE = 30 * 1000;
+
+export const DEFAULT_SUBSCRIBE_OPTIONS = {
+  keepAlive: DEFAULT_KEEP_ALIVE
+};
